chore: Reorganize and expose primitives
