+++ src/api.test.ts
@@ -1,6 +1,7 @@
 import { HTTP_METHODS } from './constants'
-import * as subject from './index'
+import * as subject from './api'
 import * as z from 'zod'
+import { HTTPMethod } from './types'
 
 export type Expect<T extends true> = T
 export type Equal<A, B> =
@@ -8,184 +9,6 @@ export type Equal<A, B> =
   (<T>() => T extends A ? 1 : 2) extends (<T>() => T extends B ? 1 : 2)
     ? true
     : false
-
-beforeEach(() => {
-  vi.clearAllMocks()
-})
-
-describe('mergeHeaders', () => {
-  it('should merge diferent kinds of Headers', () => {
-    expect(
-      subject.mergeHeaders(new Headers({ a: '1' }), { b: '2' }, [['c', '3']]),
-    ).toEqual(new Headers({ a: '1', b: '2', c: '3' }))
-  })
-
-  it('should merge diferent kinds of Headers and override values', () => {
-    expect(
-      subject.mergeHeaders(new Headers({ a: '1' }), { a: '2' }, [['a', '3']]),
-    ).toEqual(new Headers({ a: '3' }))
-  })
-
-  it('should merge diferent kinds of Headers and delete undefined values', () => {
-    expect(
-      subject.mergeHeaders(new Headers({ a: '1' }), { a: undefined }),
-    ).toEqual(new Headers({}))
-    expect(
-      subject.mergeHeaders(new Headers({ a: '1' }), { a: 'undefined' }),
-    ).toEqual(new Headers({}))
-    expect(
-      subject.mergeHeaders(new Headers({ a: '1' }), [['a', undefined]]),
-    ).toEqual(new Headers({}))
-  })
-})
-
-describe('addQueryToURL', () => {
-  it('should add the query object to a string input', () => {
-    expect(subject.addQueryToURL('https://example.com/api', { id: '1' })).toBe(
-      'https://example.com/api?id=1',
-    )
-    expect(
-      subject.addQueryToURL('https://example.com/api', 'page=2&foo=bar'),
-    ).toBe('https://example.com/api?page=2&foo=bar')
-  })
-
-  it('should add the query object to a URL input', () => {
-    expect(
-      subject.addQueryToURL(new URL('https://example.com/api'), {
-        id: '1',
-      }),
-    ).toEqual(new URL('https://example.com/api?id=1'))
-    expect(
-      subject.addQueryToURL(new URL('https://example.com/api'), 'page=2'),
-    ).toEqual(new URL('https://example.com/api?page=2'))
-  })
-
-  it('should append the query to a URL string that already has QS', () => {
-    expect(
-      subject.addQueryToURL('https://example.com/api?id=1', { page: '2' }),
-    ).toBe('https://example.com/api?id=1&page=2')
-    expect(
-      subject.addQueryToURL('https://example.com/api?id=1', 'page=2'),
-    ).toBe('https://example.com/api?id=1&page=2')
-    expect(
-      subject.addQueryToURL(
-        'https://example.com/api?id=1',
-        new URLSearchParams({ page: '2' }),
-      ),
-    ).toBe('https://example.com/api?id=1&page=2')
-  })
-
-  it('should append the query to a URL instance that already has QS', () => {
-    expect(
-      subject.addQueryToURL(new URL('https://example.com/api?id=1'), {
-        page: '2',
-      }),
-    ).toEqual(new URL('https://example.com/api?id=1&page=2'))
-    expect(
-      subject.addQueryToURL(new URL('https://example.com/api?id=1'), 'page=2'),
-    ).toEqual(new URL('https://example.com/api?id=1&page=2'))
-    expect(
-      subject.addQueryToURL(
-        new URL('https://example.com/api?id=1'),
-        new URLSearchParams({ page: '2' }),
-      ),
-    ).toEqual(new URL('https://example.com/api?id=1&page=2'))
-  })
-
-  it("should return the input in case there's no query", () => {
-    expect(subject.addQueryToURL('https://example.com/api')).toBe(
-      'https://example.com/api',
-    )
-    expect(subject.addQueryToURL(new URL('https://example.com/api'))).toEqual(
-      new URL('https://example.com/api'),
-    )
-  })
-})
-
-describe('makeGetApiURL', () => {
-  it('should return a URL which is baseURL and path joined', () => {
-    expect(subject.makeGetApiURL('https://example.com/api')('/users')).toBe(
-      'https://example.com/api/users',
-    )
-  })
-
-  it('should accept an object-like queryString and return it joined to the URL', () => {
-    const getApiURL = subject.makeGetApiURL('https://example.com/api')
-    expect(getApiURL('/users', { id: '1' })).toBe(
-      'https://example.com/api/users?id=1',
-    )
-    expect(getApiURL('/users', { active: 'true', page: '2' })).toBe(
-      'https://example.com/api/users?active=true&page=2',
-    )
-  })
-
-  it('should accept a URL as baseURL and remove extra slashes', () => {
-    expect(
-      subject.makeGetApiURL(new URL('https://example.com/api'))('/users'),
-    ).toBe('https://example.com/api/users')
-    expect(
-      subject.makeGetApiURL(new URL('https://example.com/api/'))('/users'),
-    ).toBe('https://example.com/api/users')
-    expect(
-      subject.makeGetApiURL(new URL('https://example.com/api/'))('///users'),
-    ).toBe('https://example.com/api/users')
-  })
-})
-
-describe('typedResponse', () => {
-  it('should return unknown by default when turning into a JSON', async () => {
-    const result = await subject.typedResponse(new Response('1')).json()
-    type _R = Expect<Equal<typeof result, unknown>>
-    expect(result).toEqual(1)
-  })
-
-  it('should accept a type for the JSON method', async () => {
-    const result = await subject
-      .typedResponse(new Response(`{"foo":"bar"}`))
-      .json<{ foo: string }>()
-    type _R = Expect<Equal<typeof result, { foo: string }>>
-    expect(result).toEqual({ foo: 'bar' })
-  })
-
-  it('should accept a parser for the JSON method', async () => {
-    const result = await subject
-      .typedResponse(new Response(`{"foo":"bar"}`))
-      .json(z.object({ foo: z.string() }))
-    type _R = Expect<Equal<typeof result, { foo: string }>>
-    expect(result).toEqual({ foo: 'bar' })
-  })
-})
-
-describe('ensureStringBody', () => {
-  it('should return the same if body was string', () => {
-    expect(subject.ensureStringBody('foo')).toBe('foo')
-  })
-
-  it('should return the same if body was not defined', () => {
-    expect(subject.ensureStringBody()).toBeUndefined()
-  })
-
-  it('should stringify the body if it is a JSON-like value', () => {
-    expect(subject.ensureStringBody({ page: 2 })).toBe(`{"page":2}`)
-    expect(subject.ensureStringBody([1, 2])).toBe(`[1,2]`)
-    expect(subject.ensureStringBody(3)).toBe(`3`)
-    expect(subject.ensureStringBody(true)).toBe(`true`)
-    expect(subject.ensureStringBody({})).toBe(`{}`)
-  })
-
-  it('should not stringify other valid kinds of BodyInit', () => {
-    const ab = new ArrayBuffer(0)
-    expect(subject.ensureStringBody(ab)).toBe(ab)
-    const rs = new ReadableStream()
-    expect(subject.ensureStringBody(rs)).toBe(rs)
-    const fd = new FormData()
-    expect(subject.ensureStringBody(fd)).toBe(fd)
-    const usp = new URLSearchParams()
-    expect(subject.ensureStringBody(usp)).toBe(usp)
-    const blob = new Blob()
-    expect(subject.ensureStringBody(blob)).toBe(blob)
-  })
-})
 
 const reqMock = vi.fn()
 function successfulFetch(response: string | Record<string, unknown>) {
@@ -201,6 +24,10 @@ function successfulFetch(response: string | Record<string, unknown>) {
     )
   }
 }
+
+beforeEach(() => {
+  vi.clearAllMocks()
+})
 
 describe('enhancedFetch', () => {
   describe('proxied json', () => {
@@ -462,9 +289,7 @@ describe('makeService', () => {
     const service = subject.makeService('https://example.com/api')
     for (const method of HTTP_METHODS) {
       expect(
-        typeof service[
-          method.toLocaleLowerCase() as Lowercase<subject.HTTPMethod>
-        ],
+        typeof service[method.toLocaleLowerCase() as Lowercase<HTTPMethod>],
       ).toBe('function')
     }
   })
@@ -484,5 +309,29 @@ describe('makeService', () => {
       headers: new Headers({ 'content-type': 'application/json' }),
       method: 'POST',
     })
+  })
+})
+
+describe('typedResponse', () => {
+  it('should return unknown by default when turning into a JSON', async () => {
+    const result = await subject.typedResponse(new Response('1')).json()
+    type _R = Expect<Equal<typeof result, unknown>>
+    expect(result).toEqual(1)
+  })
+
+  it('should accept a type for the JSON method', async () => {
+    const result = await subject
+      .typedResponse(new Response(`{"foo":"bar"}`))
+      .json<{ foo: string }>()
+    type _R = Expect<Equal<typeof result, { foo: string }>>
+    expect(result).toEqual({ foo: 'bar' })
+  })
+
+  it('should accept a parser for the JSON method', async () => {
+    const result = await subject
+      .typedResponse(new Response(`{"foo":"bar"}`))
+      .json(z.object({ foo: z.string() }))
+    type _R = Expect<Equal<typeof result, { foo: string }>>
+    expect(result).toEqual({ foo: 'bar' })
   })
 })
